Trying to match Logan's call in visionMain.cpp since it worked.
