--- hardware/controls/controls.h
@@ -1,6 +1,0 @@
-#ifndef CONTROLS_H
-#define CONTROLS_H
-
-void rotateAndCapture(struct Pipes pipes, float weight);
-
-#endif
