--- hardware/src/controls/README.md
@@ -1,27 +1,0 @@
-# Hardware Errata
-
-
-## rpi camera
-```plaintext
-https://www.raspberrypi.com/documentation/computers/camera_software.html
-```
-### Documentation on building own app
-```plaintext
-https://github.com/raspberrypi/documentation/blob/develop/documentation/asciidoc/computers/camera/rpicam_apps_building.adoc
-```
-
-## libcamera
-```plaintext
-https://docs.arducam.com/Raspberry-Pi-Camera/Native-camera/Libcamera-User-Guide/#introduction
-```
-### libcamera API
-```plaintext
-https://libcamera.org/api-html/namespacelibcamera.html
-```
-### simple-cam (uses libcamera from C++)
-```plaintext
-https://github.com/kbingham/simple-cam/blob/master/simple-cam.cpp
-```
-```plaintext
-
-```
